refactor(ui): hoist provider negotiation into shared papi package
The SignedTerms/NegotiateRequest types and the negotiateTerms() /negotiate
client were byte-identical copies in both s3-client and drive-client, and the
NewBucketDialog/NewDriveDialog provider-select handlers shared the same
multiaddr-parse + negotiate + error-mapping block.

Hoist all of it into @web3-storage/papi:
- move SignedTerms, NegotiateRequest, negotiateTerms there (single source of
  truth); each client lib now re-exports SignedTerms so state modules importing
  it from the client facade are unaffected.
- add negotiateProviderTerms(provider, request), which resolves the multiaddr
  to a URL and negotiates in one call, returning a discriminated
  { ok, url, signed } | { ok: false, error } result. Both dialogs collapse to a
  single call + result check, dropping the nested try/catch and the duplicated
  unparseable-multiaddr / negotiate-failure messages.

Net -19 lines; behaviour unchanged.

Verified on toaster: tsc -b && vite build pass for drive-ui and s3-ui; drive-ui vitest 18/18 pass.
